move sqlite registration warning to cataloger creation
