+++ syft/pkg/cataloger/rpm/parse_rpm_db.go
@@ -1,7 +1,6 @@
 package rpm
 
 import (
-	"database/sql"
 	"fmt"
 	"io"
 	"os"
@@ -20,11 +19,6 @@ import (
 
 // parseRpmDb parses an "Packages" RPM DB and returns the Packages listed within it.
 func parseRpmDB(resolver source.FileResolver, env *generic.Environment, reader source.LocationReadCloser) ([]pkg.Package, []artifact.Relationship, error) {
-	// check if a sqlite driver is available
-	if !isSqliteDriverAvailable() {
-		log.Warnf("sqlite driver is not available, newer RPM databases might not be cataloged")
-	}
-
 	f, err := os.CreateTemp("", internal.ApplicationName+"-rpmdb")
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to create temp rpmdb file: %w", err)
@@ -123,9 +117,4 @@ func extractRpmdbFileRecords(resolver source.FilePathResolver, entry rpmdb.Packa
 		}
 	}
 	return records
-}
-
-func isSqliteDriverAvailable() bool {
-	_, err := sql.Open("sqlite", ":memory:")
-	return err == nil
 }
